updated candidate setup with utils
